Add files via upload
This is a diagram I created for a previous project I had been working on until it was canceled.  It lists some of the sensors available in Android, and whether they need to be manually polled for new data, or you request Android to give a stream/events as new data becomes available.  It's just meant to be a reference, I thought some of the information might be interesting.